Update frame calculation for string list to dynamically support overlapping titles and values.
In most cases we want to give the title's content precedence over the value, such that the value will scale down in the available space, and then truncate. However, we don't want a long title to be able to leave no room at all for the value, so for a long title, if the value is also long enough that they won't both fit, the limit the title to a fixed percentage of the cell's width, so that there's still room for the remaining portion to be used by the value.