Simplify internal interface (#146)
Way back when we started this project, we thought that we'd want to have "Tweak backups" - where you could create a backup of your tweaks, then go off and experiment with different animation settings, then save/restore/share/import different configurations. This would be helpful for teammates to share different animation configurations, or experiment with their tweaks without losing their current place. Basically, it'd kinda be like version-control but for your tweaks?

Anyways, in the many years of this project, this feature has still not been made - so rather than keep vestigial, unused code around, let's just drop it and simplify the UI code of SwiftTweaks.

Now, TweaksRootViewController has been dropped, and the TweaksBackupViewController is gone, too. The view-controller hierarchy is now:

TweaksViewController (the root view controller that is first shown, handles creation of the navigation-controller and the TweaksCollectionListViewController)
TweaksCollectionListViewController (the list-of-collections screen; the root-level contents of the Tweaks menu!)
TweaksCollectionViewController (which shows tweak groups within a given collection)
(occasionally) other child view controllers, e.g. choosing from a StringOption, or editing a Color.